requirements.txt
django==5.2.3
djangorestframework==3.16.0
whitenoise==6.6.0
python-decouple==3.8
django-cors-headers==4.7.0
mysqlclient==2.2.7